# backend/openapi.json
{
  "components": {
    "schemas": {
      "Category": {
        "enum": [
          "water",
          "electricity",
          "sanitation",
          "roads",
          "streetlights",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
      "ComplaintCreate": {
        "additionalProperties": false,
        "examples": [
          {
            "location": "Street 12, Block 4, Gulshan-e-Iqbal, Karachi",
            "text": "Burst water main flooding Street 12 since fajr, water entering ground floors"
          }
        ],
        "properties": {
          "location": {
            "maxLength": 200,
            "minLength": 3,
            "title": "Location",
            "type": "string"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Contact"
          },
          "text": {
            "description": "What is wrong",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "location"
        ],
        "title": "ComplaintCreate",
        "type": "object"
      },
      "ComplaintCreated": {
        "properties": {
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "allowed_transitions": {
            "description": "Statuses this complaint may move to next, decided by the server's state machine",
            "items": {
              "$ref": "#/components/schemas/Status"
            },
            "title": "Allowed Transitions",
            "type": "array"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Contact"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "triage": {
            "$ref": "#/components/schemas/TriageInfo"
          },
          "triage_latency_ms": {
            "title": "Triage Latency Ms",
            "type": "integer"
          },
          "triaged_by": {
            "$ref": "#/components/schemas/TriagedBy"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "location",
          "reporter_contact",
          "category",
          "priority",
          "status",
          "ai_summary",
          "triaged_by",
          "triage_latency_ms",
          "created_at",
          "updated_at",
          "allowed_transitions",
          "triage"
        ],
        "title": "ComplaintCreated",
        "type": "object"
      },
      "ComplaintOut": {
        "properties": {
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "allowed_transitions": {
            "description": "Statuses this complaint may move to next, decided by the server's state machine",
            "items": {
              "$ref": "#/components/schemas/Status"
            },
            "title": "Allowed Transitions",
            "type": "array"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Contact"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "triage_latency_ms": {
            "title": "Triage Latency Ms",
            "type": "integer"
          },
          "triaged_by": {
            "$ref": "#/components/schemas/TriagedBy"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "location",
          "reporter_contact",
          "category",
          "priority",
          "status",
          "ai_summary",
          "triaged_by",
          "triage_latency_ms",
          "created_at",
          "updated_at",
          "allowed_transitions"
        ],
        "title": "ComplaintOut",
        "type": "object"
      },
      "ComplaintPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComplaintOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "ComplaintPage",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "FieldErrorOut": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "FieldErrorOut",
        "type": "object"
      },
      "HealthOut": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthOut",
        "type": "object"
      },
      "Priority": {
        "enum": [
          "high",
          "normal",
          "low"
        ],
        "title": "Priority",
        "type": "string"
      },
      "ProvidersOut": {
        "properties": {
          "active": {
            "description": "triaged_by value the active provider writes, e.g. llm:gemini",
            "title": "Active",
            "type": "string"
          },
          "cache": {
            "$ref": "#/components/schemas/TriageCacheOut"
          },
          "configured": {
            "description": "TRIAGE_PROVIDER setting",
            "title": "Configured",
            "type": "string"
          },
          "fallback": {
            "title": "Fallback",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "recent": {
            "items": {
              "$ref": "#/components/schemas/TriageOutcomeOut"
            },
            "title": "Recent",
            "type": "array"
          },
          "timeout_s": {
            "title": "Timeout S",
            "type": "number"
          }
        },
        "required": [
          "active",
          "configured",
          "fallback",
          "model",
          "timeout_s",
          "cache",
          "recent"
        ],
        "title": "ProvidersOut",
        "type": "object"
      },
      "RateLimitErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retry_after_s": {
            "title": "Retry After S",
            "type": "integer"
          }
        },
        "required": [
          "detail",
          "retry_after_s"
        ],
        "title": "RateLimitErrorOut",
        "type": "object"
      },
      "ReadinessOut": {
        "properties": {
          "checks": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Checks",
            "type": "object"
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "title": "Failed",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks",
          "failed"
        ],
        "title": "ReadinessOut",
        "type": "object"
      },
      "StatsOut": {
        "properties": {
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Category"
            },
            "title": "By Category",
            "type": "object"
          },
          "by_priority": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Priority"
            },
            "title": "By Priority",
            "type": "object"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Status"
            },
            "title": "By Status",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "by_category",
          "by_priority",
          "by_status"
        ],
        "title": "StatsOut",
        "type": "object"
      },
      "Status": {
        "enum": [
          "open",
          "in_progress",
          "resolved",
          "rejected"
        ],
        "title": "Status",
        "type": "string"
      },
      "StatusUpdate": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusUpdate",
        "type": "object"
      },
      "TransitionErrorOut": {
        "properties": {
          "allowed": {
            "items": {
              "$ref": "#/components/schemas/Status"
            },
            "title": "Allowed",
            "type": "array"
          },
          "attempted": {
            "$ref": "#/components/schemas/Status"
          },
          "current": {
            "$ref": "#/components/schemas/Status"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "detail",
          "current",
          "attempted",
          "allowed"
        ],
        "title": "TransitionErrorOut",
        "type": "object"
      },
      "TriageCacheOut": {
        "properties": {
          "hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit Rate"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "misses": {
            "title": "Misses",
            "type": "integer"
          }
        },
        "required": [
          "hits",
          "misses",
          "hit_rate"
        ],
        "title": "TriageCacheOut",
        "type": "object"
      },
      "TriageInfo": {
        "properties": {
          "cache_hit": {
            "title": "Cache Hit",
            "type": "boolean"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "fallback": {
            "title": "Fallback",
            "type": "boolean"
          }
        },
        "required": [
          "confidence",
          "cache_hit",
          "fallback"
        ],
        "title": "TriageInfo",
        "type": "object"
      },
      "TriageOutcomeOut": {
        "properties": {
          "at": {
            "title": "At",
            "type": "string"
          },
          "cache_hit": {
            "title": "Cache Hit",
            "type": "boolean"
          },
          "complaint_id": {
            "title": "Complaint Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "fallback": {
            "title": "Fallback",
            "type": "boolean"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "complaint_id",
          "provider",
          "latency_ms",
          "fallback",
          "cache_hit",
          "error",
          "at"
        ],
        "title": "TriageOutcomeOut",
        "type": "object"
      },
      "TriagedBy": {
        "description": "Which component produced the triage result stored on a complaint.",
        "enum": [
          "llm:gemini",
          "llm:groq",
          "llm:ollama",
          "rules",
          "rules:fallback",
          "simulated"
        ],
        "title": "TriagedBy",
        "type": "string"
      },
      "ValidationErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldErrorOut"
            },
            "title": "Errors",
            "type": "array"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "detail",
          "errors"
        ],
        "title": "ValidationErrorOut",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Municipal complaint intake, AI triage and operations.",
    "title": "CivicPulse API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/complaints": {
      "get": {
        "operationId": "list_complaints_api_complaints_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Status"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "List Complaints",
        "tags": [
          "complaints"
        ]
      },
      "post": {
        "operationId": "submit_complaint_api_complaints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Field-level validation errors"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded; see Retry-After"
          }
        },
        "summary": "Submit Complaint",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{complaint_id}": {
      "get": {
        "operationId": "get_complaint_api_complaints__complaint_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Get Complaint",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{complaint_id}/status": {
      "patch": {
        "operationId": "change_status_api_complaints__complaint_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionErrorOut"
                }
              }
            },
            "description": "Transition not allowed"
          }
        },
        "summary": "Change Status",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/meta/providers": {
      "get": {
        "operationId": "get_providers_api_meta_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Providers",
        "tags": [
          "stats",
          "meta"
        ]
      }
    },
    "/api/stats": {
      "get": {
        "operationId": "get_stats_api_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Cache": {
                "description": "HIT or MISS",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Stats",
        "tags": [
          "stats"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "ops"
        ]
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessOut"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ready",
        "tags": [
          "ops"
        ]
      }
    }
  }
}